# Add xtal frequency configuration