--- a/esp-hal/src/clock/mod.rs
+++ b/esp-hal/src/clock/mod.rs
@@ -341,7 +341,7 @@ impl Clocks {
         } else {
             const {
                 match esp_config::esp_config_str!("ESP_HAL_CONFIG_XTAL_FREQUENCY").as_bytes() {
-                    b"auto" => XtalClock::Other(0),
+                    b"auto" => XtalClock::Other(0), // Can't be `unreachable!` due to const eval.
                     b"26" => XtalClock::_26M,
                     b"40" => XtalClock::_40M,
                     other => XtalClock::Other(esp_config::esp_config_int_parse!(u32, other)),
@@ -392,7 +392,7 @@ impl Clocks {
         } else {
             const {
                 match esp_config::esp_config_str!("ESP_HAL_CONFIG_XTAL_FREQUENCY").as_bytes() {
-                    b"auto" => XtalClock::Other(0),
+                    b"auto" => XtalClock::Other(0), // Can't be `unreachable!` due to const eval.
                     b"26" => XtalClock::_26M,
                     b"40" => XtalClock::_40M,
                     other => XtalClock::Other(esp_config::esp_config_int_parse!(u32, other)),