This PR adds `ESP_HAL_CONFIG_XTAL_FREQUENCY` - although only 32 and C2 have meaningful configurability. THe point of this PR is to allow const-propagating `xtal_freq` wherever possible, mostly affecting `esp_hal::time::now()` and indirectly embassy-time.

We could probably double-check in `Clocks::init` that the configured frequency matches what we measure, I don't know how badly we would want that, though.

---

Running a modified `embassy_executor_benchmark` on the S2 (no time queue tasks, Task1 reading current time in every loop):

| Test case | main | PR | 
| --------- | ---- | -- |
| no-op     | count=483798, cycles=24803/100 | count=483791, cycles=24804/100 |
| `now()`   | count=357079, cycles=33606/100 | count=373770, cycles=32105/100 |
| `black_box(now())` | count=279690, cycles=42904/100 | count=330518, cycles=36306/100 |

- count: iterations in 500ms, higher is better
- cycles: CPU clock cycles per iteration (x100), lower is better